pyproject.toml
[project]
name = "michael-levin-qa-engine"
version = "0.1.0"
description = "A RAG system for Michael Levin's research papers and publications"
authors = [
    {name = "Real Pattern Laboratory", email = "real.pattern.laboratory@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["research", "digital-twin", "scraping", "pdf", "embeddings"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "streamlit>=1.28.0",
    "openai>=1.0.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
    "python-dotenv>=1.0.0",
    "pypdf>=3.0.0",
    "pdfplumber>=0.9.0",
    "PyMuPDF>=1.23.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "google-generativeai>=0.3.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
    "toml>=0.10.0",
    "tiktoken>=0.10.0",
    "assemblyai>=0.42.1",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.scripts]
michael-levin-qa = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "requests.*",
    "bs4.*",
    "PyPDF2.*",
    "pdfplumber.*",
    "sentence_transformers.*",
    "nltk.*",
    "chromadb.*",
    "faiss.*",
    "openai.*",
    "transformers.*",
    "tqdm.*",
]
ignore_missing_imports = true 
